refactor: Use same renderer for chart preview/published
